Consolidate metadata utilities (explore.services, lib/field_type.js) into schema_metadata.js
