--- resources/frontend_client/app/explore/explore.module.js
@@ -1,6 +1,0 @@
-'use strict';
-
-// Explore (Metabase)
-angular.module('metabase.explore', [
-    'metabase.explore.services'
-]);
